Don't update last ping time on data received
Thanks to Nam Nguyen for reporting this issue.

Quoting from https://github.com/1stvamp/hippybot/issues/2:

Apparently the problem is with JabberBot's handling of its private
__lastping variable. It resets this variable when it receives a packet
from the server. I suppose this might be reasonable with other jabberd
but, as explained by @powdahound, Hipchat only counts the packets that
are sent from client, not packets that Hipchat itself has sent. So there
might be times when the if statement will fail because time.time() -
self.__JabberBot_lastping is less than PING_FREQUENCY.